Extract shared transit network dim and route vehicle fetch
